--- src/main/java/com/hhplusconcert/concert/usecase/CreateQueueUseCase.java
@@ -1,19 +1,0 @@
-package com.hhplusconcert.concert.usecase;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-public class CreateQueueUseCase {
-
-    @Getter
-    @AllArgsConstructor
-    public static class Input {
-        Long userId;
-    }
-
-    @AllArgsConstructor
-    @Getter
-    public static class Output {
-        String token;
-    }
-}
